refactor(emit): split oversized emitter files to satisfy 2000-line cap
`source_file/mod.rs` crossed 2000 lines in this branch; extract tests
to `es5_emit_tests.rs` (1990 lines).

`ir_printer.rs` already exceeded 2000 on main; extract namespace/enum/
block emission helpers (~385 lines) to `ir_printer_namespace.rs` (395
lines) so the changed `emit_namespace_iife` lives in a compliant file.

`class_es5_ir.rs` already exceeded 2000 on main; extract constructor
and super-call emit methods (~1343 lines) to
`class_es5_ir_constructor.rs` (1366 lines) so the changed
`derived_constructor_needs_pre_super_this_capture` lives in a
compliant file.

All new files are under 2000 lines. `cargo clippy -D warnings` and
all previously-passing unit tests remain green.
